fix(*): move data recieving to AppContent component
