+++ src/components/burger-ingredients/types.ts
@@ -28,3 +28,8 @@ export type Ingredient = {
 export type IngredientWithCounter = Ingredient & {
     quantity?: number;
 };
+
+export type BurgerIngredientsProps = {
+    // FIXME убрать undefined после проработки запроса за данными
+    ingredients: Map<string, Array<IngredientWithCounter>> | undefined;
+};
